+++ src/main/java/it/gov/pagopa/pu/bff/connector/auth/config/AuthApisHolder.java
@@ -4,7 +4,7 @@ import it.gov.pagopa.pu.auth.controller.ApiClient;
 import it.gov.pagopa.pu.auth.controller.BaseApi;
 import it.gov.pagopa.pu.auth.controller.generated.AuthnApi;
 import it.gov.pagopa.pu.auth.controller.generated.AuthzApi;
-import it.gov.pagopa.pu.bff.config.RestTemplateConfig;
+import it.gov.pagopa.pu.bff.config.rest.RestTemplateConfig;
 import jakarta.annotation.PreDestroy;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Service;
