+++ src/main/java/it/gov/pagopa/pu/bff/util/HttpUtils.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.bff.util;
 
-import it.gov.pagopa.pu.bff.config.HttpClientConfig;
+import it.gov.pagopa.pu.bff.config.rest.HttpClientConfig;
 import org.apache.hc.client5.http.config.ConnectionConfig;
 import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManagerBuilder;
 import org.apache.hc.client5.http.ssl.TlsSocketStrategy;
